+++ README/CHANGELOG.md
@@ -379,16 +379,18 @@ I tested it on Android 12 & 13 on two devices.
 - Porn Fetch supports discord rich presence (although it's optional and disabled by default)
 - Porn Fetch now checks connection to all sites and lets you know if there was an error
 - Untested support for macOS
+- Added model support for xvideos
 
 ## Bug fixes
 - Fixed application tabbing
-- 
+
 ## Other
 - Huge code refactoring
 - Added CI/CD build actions for all platforms
 - Switched to Qt 6.7.0
 - Improved terminal debug messages
 - Reduced file size by over 70% on Windows and Linux
+- Improved the tree widget
 
 ## Deprecations
 - removed all metadata functionality from Porn Fetch, because it's useless and hard to maintain.